requirements.txt
customtkinter>=5.2.0
yt-dlp>=2024.1.0
Pillow>=10.0.0
mutagen>=1.47.0
